+++ tests/load/load_test.go
@@ -12,39 +12,20 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ava-labs/avalanchego/api/health"
 	"github.com/ava-labs/subnet-evm/tests/utils"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/go-cmd/cmd"
 	ginkgo "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 )
 
-var startCmd *cmd.Cmd
+func init() {
+	utils.RegisterNodeRun()
+}
 
 func TestE2E(t *testing.T) {
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	ginkgo.RunSpecs(t, "subnet-evm small load simulator test suite")
 }
-
-// BeforeSuite starts an AvalancheGo process to use for the e2e tests
-var _ = ginkgo.BeforeSuite(func() {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
-	wd, err := os.Getwd()
-	gomega.Expect(err).Should(gomega.BeNil())
-	log.Info("Starting AvalancheGo node", "wd", wd)
-	startCmd, err = utils.RunCommand("./scripts/run.sh")
-	gomega.Expect(err).Should(gomega.BeNil())
-
-	// Assumes that startCmd will launch a node with HTTP Port at [utils.DefaultLocalNodeURI]
-	healthClient := health.NewClient(utils.DefaultLocalNodeURI)
-	healthy, err := health.AwaitReady(ctx, healthClient, 5*time.Second, nil)
-	gomega.Expect(err).Should(gomega.BeNil())
-	gomega.Expect(healthy).Should(gomega.BeTrue())
-	log.Info("AvalancheGo node is healthy")
-})
 
 var _ = ginkgo.Describe("[Load Simulator]", ginkgo.Ordered, func() {
 	ginkgo.It("basic subnet load test", ginkgo.Label("load"), func() {
@@ -69,11 +50,4 @@ var _ = ginkgo.Describe("[Load Simulator]", ginkgo.Ordered, func() {
 		fmt.Printf("\nCombined output:\n\n%s\n", string(out))
 		gomega.Expect(err).Should(gomega.BeNil())
 	})
-})
-
-var _ = ginkgo.AfterSuite(func() {
-	gomega.Expect(startCmd).ShouldNot(gomega.BeNil())
-	gomega.Expect(startCmd.Stop()).Should(gomega.BeNil())
-	// TODO add a new node to bootstrap off of the existing node and ensure it can bootstrap all subnets
-	// created during the test
 })
