+++ tests/precompile/solidity/suites.go
@@ -8,20 +8,13 @@ import (
 	"context"
 	"time"
 
-	"github.com/ava-labs/avalanchego/api/health"
 	"github.com/ava-labs/subnet-evm/tests/utils"
 	ginkgo "github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
 )
 
-var _ = ginkgo.Describe("[Precompiles]", ginkgo.Ordered, func() {
-	ginkgo.It("ping the network", ginkgo.Label("setup"), func() {
-		client := health.NewClient(utils.DefaultLocalNodeURI)
-		healthy, err := client.Readiness(context.Background(), nil)
-		gomega.Expect(err).Should(gomega.BeNil())
-		gomega.Expect(healthy.Healthy).Should(gomega.BeTrue())
-	})
-})
+func init() {
+	utils.RegisterPingTest()
+}
 
 var _ = ginkgo.Describe("[Precompiles]", ginkgo.Ordered, func() {
 	// Each ginkgo It node specifies the name of the genesis file (in ./tests/precompile/genesis/)
