+++ tests/utils/command.go
@@ -4,12 +4,17 @@
 package utils
 
 import (
+	"context"
 	"fmt"
+	"os"
 	"strings"
 	"time"
 
+	"github.com/ava-labs/avalanchego/api/health"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/go-cmd/cmd"
+	"github.com/onsi/ginkgo/v2"
+	"github.com/onsi/gomega"
 )
 
 // RunCommand starts the command [bin] with the given [args] and returns the command to the caller
@@ -42,4 +47,47 @@ func RunCommand(bin string, args ...string) (*cmd.Cmd, error) {
 	}()
 
 	return curCmd, nil
+}
+
+func RegisterPingTest() {
+	ginkgo.Describe("[Precompiles]", ginkgo.Ordered, func() {
+		ginkgo.It("ping the network", ginkgo.Label("ping"), func() {
+			client := health.NewClient(DefaultLocalNodeURI)
+			healthy, err := client.Readiness(context.Background(), nil)
+			gomega.Expect(err).Should(gomega.BeNil())
+			gomega.Expect(healthy.Healthy).Should(gomega.BeTrue())
+		})
+	})
+}
+
+// RegisterNodeRun registers a before suite that starts an AvalancheGo process to use for the e2e tests
+// and an after suite that stops the AvalancheGo process
+func RegisterNodeRun() {
+	// BeforeSuite starts an AvalancheGo process to use for the e2e tests
+	var startCmd *cmd.Cmd
+	_ = ginkgo.BeforeSuite(func() {
+		ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+		defer cancel()
+
+		wd, err := os.Getwd()
+		gomega.Expect(err).Should(gomega.BeNil())
+		log.Info("Starting AvalancheGo node", "wd", wd)
+		cmd, err := RunCommand("./scripts/run.sh")
+		startCmd = cmd
+		gomega.Expect(err).Should(gomega.BeNil())
+
+		// Assumes that startCmd will launch a node with HTTP Port at [utils.DefaultLocalNodeURI]
+		healthClient := health.NewClient(DefaultLocalNodeURI)
+		healthy, err := health.AwaitReady(ctx, healthClient, 5*time.Second, nil)
+		gomega.Expect(err).Should(gomega.BeNil())
+		gomega.Expect(healthy).Should(gomega.BeTrue())
+		log.Info("AvalancheGo node is healthy")
+	})
+
+	ginkgo.AfterSuite(func() {
+		gomega.Expect(startCmd).ShouldNot(gomega.BeNil())
+		gomega.Expect(startCmd.Stop()).Should(gomega.BeNil())
+		// TODO add a new node to bootstrap off of the existing node and ensure it can bootstrap all subnets
+		// created during the test
+	})
 }
